requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pydantic>=2.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pysrt>=1.1.2
chardet>=5.0.0
spacy>=3.5.0
language-tool-python>=2.7.0
transformers>=4.25.0
torch>=2.0.0
nltk>=3.8
jinja2>=3.1.2
websockets>=11.0.0
pydub>=0.25.1
pytest>=7.3.1
requests>=2.31.0
opencv-python>=4.7.0
numpy>=1.24.0
